% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CpGs_select.R
\name{CpGs_select}
\alias{CpGs_select}
\title{extract data from the list of GSE objects}
\usage{
CpGs_select(mergedData, top_CpGs_number = 20)
}
\arguments{
\item{mergedData}{the result of the MetDataExtracted}

\item{top_CpGs_number}{numeric}
}
\value{
character vector of the probeID
}
\description{
to extract β value matrix and age information from GSE objects.
then organize these data, preparing them for downstream analysis
the column name of age in pheno_data is not certain, so it is recommended to check the annotation file first.
we only consider "age:ch1","age (y):ch1" these two names
though dataset_names should be omitted, I do not make it for my laziness :).
}
\examples{
CpGs_select(mergedData,top_CpGs_number = 20)

}
