% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MetDataDownload.R
\name{MetDataDownload}
\alias{MetDataDownload}
\title{download illumina 450k methylation data from GEO}
\usage{
MetDataDownload(datasets_names)
}
\arguments{
\item{dataset_names}{character vector}
}
\value{
list of GSE objects
}
\description{
to download data from GEO database, it may take a long time due to the large amount of data.
also, in some versions of GEOquery packages, the getGEO function only allow for download in 2 minutes.
in that case, it is  recommended to download data by hands or update the version of GEOquery.
}
\examples{
MetDataDownload(c('GSE40279','GSE67705'))

}
