% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MetDataMerge.R
\name{MetDataMerge}
\alias{MetDataMerge}
\title{merge all the samples}
\usage{
MetDataMerge(all_list)
}
\arguments{
\item{all_list}{the list got from MetDataExtract}

\item{dataset_names}{character vector}
}
\value{
a dataframe ready for model training
}
\description{
merge all the samples into one dataframe
}
\examples{
MetDataMerge(all_list,datasets_names)
}
