% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MetDataModelEvaluation.R
\name{MetDataModelEvaluation}
\alias{MetDataModelEvaluation}
\title{Model evaluation}
\usage{
MetDataModelEvaluation(glmnet.Training, PartitionedData)
}
\arguments{
\item{model}{model constructed in the last step}

\item{beta_matrix_train}{beta value matrix}

\item{age_train}{numeric vector}

\item{beta_matrix_test}{beta value matrix}

\item{age_test}{numeric vector}
}
\value{
results list
}
\description{
predict age with the model constructed
}
\examples{
MetDataModelEvaluation(glmnet.Training,beta_matrix_test,beta_matrix_train,trainSet,testSet)

}
