% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MetDataMultiRegression.R
\name{MetDataMultiRegression}
\alias{MetDataMultiRegression}
\title{multiple-variate regression}
\usage{
MetDataMultiRegression(PartitionedData)
}
\arguments{
\item{PartitionedData}{list}
}
\value{
MultiRegressionmodel
}
\description{
to build a multiple-variate regression model
}
\examples{
MetDataMultiRegression(PartitionedData)


}
