% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MetDataPartition.R
\name{MetDataPartition}
\alias{MetDataPartition}
\title{divide samples for trainSet and testSet}
\usage{
MetDataPartition(dat, trainProportion = 0.8)
}
\arguments{
\item{dat}{the dataframe as a result of MetDataMerge}
}
\value{
the four parts contain beta_matrix_train,beta_matrix_test,age_train,age_test
}
\description{
since the data is ready for the model construction, now we prepare the trainSet and testSet
}
\examples{
MetDataPartition(dat)
}
