% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MetDataTrainElasticNet.R
\name{MetDataTrainElasticNet}
\alias{MetDataTrainElasticNet}
\title{elastic net model construction}
\usage{
MetDataTrainElasticNet(PartitionedData, para_alpha = 0.5)
}
\arguments{
\item{beta_matrix_train}{beta value matrix}

\item{age_train}{numeric vector}
}
\value{
elastic net model
}
\description{
build a elastic net by glmnet, the default alpha value is 0.5
}
\examples{
MetDataTrainElasticNet(beta_matrix_train, age_train,para_alpha = 0.5)

}
