% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MetDataTrainLasso.R
\name{MetDataTrainLasso}
\alias{MetDataTrainLasso}
\title{lasso model construction}
\usage{
MetDataTrainLasso(PartitionedData)
}
\arguments{
\item{beta_matrix_train}{beta value matrix}

\item{age_train}{numeric vector}
}
\value{
lasso model
}
\description{
build a lasso model by glmnet
}
\examples{
MetDataTrainLasso(beta_matrix_train, age_train,para_alpha = 1)

}
